Add files via upload
Title: International Time Series Modeling Competition - Top 6 Submission

Description:
Welcome to the GitHub repository for our project, "Multivariate Share Price Forecasting of SBI and HDFC Banks ." This repository contains the culmination of our hard work as a team of three statistics students from Mangalore University. Our project was undertaken as part of an international time series modeling competition held in March-April 2023.

Project Overview:
In this repository, you will find the main deliverable of our project - the Manuscript PDF. This document encapsulates our research, analysis, findings, and insights into the time series modeling task we were assigned as part of the competition. Our project revolves around the exploration and application of statistical techniques to tackle complex time series data.

Achievements:
We are proud to share that our submission was selected as one of the Top 6 projects out of over 60 submissions from around the world. This recognition highlights the rigor and quality of our work in the field of time series modeling. As part of our journey in the competition, we had the honor of delivering a final presentation on this project, showcasing our methodology and results to a wider audience.